F64 to f32

+++ src/main.rs
@@ -12,7 +12,7 @@ use vec3::Vec3;
 use Vec3 as Color;
 use Vec3 as Point3;
 
-pub fn hit_sphere(center: &Point3, radius: f64, ray: &Ray) -> f64 {
+pub fn hit_sphere(center: &Point3, radius: f32, ray: &Ray) -> f32 {
     let co = *center - ray.origin();
     let a = &ray.direction().length_squared();
     let h = Vec3::dot(&ray.direction(), &co);
@@ -58,7 +58,7 @@ pub fn calculate_chank(
     for j in y_start..y_end {
         for i in x_start..x_end {
             let pixel_center =
-                pixel00_loc + (i as f64 * pixel_delta_u) + (j as f64 * pixel_delta_v);
+                pixel00_loc + (i as f32 * pixel_delta_u) + (j as f32 * pixel_delta_v);
 
             let ray_direction = pixel_center - camera_center;
             let r = Ray::new(camera_center, ray_direction);
@@ -78,14 +78,14 @@ fn main() {
 
     // Calculate the image height, and ensure that it's at least 1.
 
-    let mut image_height: i32 = (image_width as f64 / aspect_ratio) as i32;
+    let mut image_height: i32 = (image_width as f32 / aspect_ratio) as i32;
     image_height = if image_height < 1 { 1 } else { image_height };
 
     // Camera
 
     let focal_length = 1.0;
     let viewport_height = 2.0;
-    let viewport_width = viewport_height * (image_width as f64 / image_height as f64);
+    let viewport_width = viewport_height * (image_width as f32 / image_height as f32);
     let camera_center = Point3::zero();
 
     // Calculate the vectors across the horizontal and down the vertical viewport edges.
@@ -95,8 +95,8 @@ fn main() {
 
     // Calculate the horizontal and vertical delta vectors from pixel to pixel.
 
-    let pixel_delta_u = viewport_u / image_width as f64;
-    let pixel_delta_v = viewport_v / image_height as f64;
+    let pixel_delta_u = viewport_u / image_width as f32;
+    let pixel_delta_v = viewport_v / image_height as f32;
 
     // Calculate the location of the upper left pixel.
 
@@ -112,8 +112,7 @@ fn main() {
 
     let start_time = Instant::now();
 
-    let chunks_x = 1; // Right now to lazy to implement horizontal chunking
-                      // If I keep using ppm, I need to concatenate lines from each chunk in proper order to output correct image. Already significant boost received, though seems a bit weird, that it dropped from 12 sec to 178 ms for such monotonous calculation, would expect to be 6X increase as now it is done in parallel on many cores, but not 100X
+    let chunks_x = 1; // Right now to lazy to implement horizontal chunking If I keep using ppm, I need to concatenate lines from each chunk in proper order to output correct image.
 
     let chunks_y = (image_height + chunk_dim - 1) / chunk_dim;
