+++ src/ray.rs
@@ -23,7 +23,7 @@ impl Ray {
         self.dir
     }
 
-    pub fn at(&self, t: f64) -> Point3 {
+    pub fn at(&self, t: f32) -> Point3 {
         self.orig + self.dir * t
     }
 }
